+++ R/kinfitr_bloodmodels.R
@@ -725,8 +725,9 @@ blmod_exp_startpars <- function(time, activity, fit_exp3=TRUE,
 #' @param multstart_upper Optional. The upper limits of the starting parameters.
 #' @param multstart_iter The number of fits to perform with different starting
 #'   parameters. If set to 1, then the starting parameters will be used for a
-#'   single fit. Otherwise the starting values are chosen by improved Latin
-#'   hypercube sampling over the multstart bounds rather than at random.
+#'   single fit. Otherwise a single number selects the starting values by
+#'   improved Latin hypercube sampling over the multstart bounds rather than at
+#'   random, while one value per parameter forms a Cartesian grid.
 #' @param Method_weights If no weights provided, should the weights be divided
 #'   by discrete and continuous samples equally (i.e. with more continuous
 #'   samples, the continuous samples each get less weight). Default is TRUE.
@@ -1630,8 +1631,9 @@ blmod_feng_startpars <- function(time, activity,
 #' @param multstart_upper Optional. The upper limits of the starting parameters.
 #' @param multstart_iter The number of fits to perform with different starting
 #'   parameters. If set to 1, then the starting parameters will be used for a
-#'   single fit. Otherwise the starting values are chosen by improved Latin
-#'   hypercube sampling over the multstart bounds rather than at random.
+#'   single fit. Otherwise a single number selects the starting values by
+#'   improved Latin hypercube sampling over the multstart bounds rather than at
+#'   random, while one value per parameter forms a Cartesian grid.
 #' @param Method_weights If no weights provided, should the weights be divided
 #'   by discrete and continuous samples equally (i.e. with more continuous
 #'   samples, the continuous samples each get less weight). Default is TRUE.
@@ -1963,8 +1965,9 @@ predict_blood_feng <- function(object, newdata = NULL) {
 #' @param multstart_upper Optional. The upper limits of the starting parameters.
 #' @param multstart_iter The number of fits to perform with different starting
 #'   parameters. If set to 1, then the starting parameters will be used for a
-#'   single fit. Otherwise the starting values are chosen by improved Latin
-#'   hypercube sampling over the multstart bounds rather than at random.
+#'   single fit. Otherwise a single number selects the starting values by
+#'   improved Latin hypercube sampling over the multstart bounds rather than at
+#'   random, while one value per parameter forms a Cartesian grid.
 #' @param Method_weights If no weights provided, should the weights be divided
 #'   by discrete and continuous samples equally (i.e. with more continuous
 #'   samples, the continuous samples each get less weight). Default is TRUE.
@@ -2379,8 +2382,9 @@ predict_blood_fengconv <- function(object, newdata = NULL) {
 #' @param multstart_upper Optional. The upper limits of the starting parameters.
 #' @param multstart_iter The number of fits to perform with different starting
 #'   parameters. If set to 1, then the starting parameters will be used for a
-#'   single fit. Otherwise the starting values are chosen by improved Latin
-#'   hypercube sampling over the multstart bounds rather than at random.
+#'   single fit. Otherwise a single number selects the starting values by
+#'   improved Latin hypercube sampling over the multstart bounds rather than at
+#'   random, while one value per parameter forms a Cartesian grid.
 #' @param Method_weights If no weights provided, should the weights be divided
 #'   by discrete and continuous samples equally (i.e. with more continuous
 #'   samples, the continuous samples each get less weight). Default is TRUE.